backend/user/requirements.txt
Flask==2.3.3
supabase==2.13.0
python-dotenv==0.15.0
requests==2.32.3
<<<<<<< Updated upstream
pika==1.3.2
gunicorn==21.2.0
=======
gunicorn==21.2.0
Flask_Cors==5.0.0
>>>>>>> Stashed changes
